결제 및 포인트 관련 중복 어노테이션 제거 및 type 변경(엔티티:int -> long, DB:int -> bigint), created_by 크기 확장varchar(20->45)

+++ src/main/java/com/project/momo/entity/AdjustPoint.java
@@ -5,7 +5,6 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
 
 @Entity
 @Table(name = "ADJUST_POINTS")
@@ -17,7 +16,7 @@ public class AdjustPoint extends BaseEntity {
     private Long id;
 
     @Min(1)
-    private int amount;
+    private long amount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "memebr_id")
